loader/requirements.txt
argcomplete==3.5.1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
commitizen==3.29.1
cryptography==43.0.3
decli==0.6.2
distlib==0.3.9
filelock==3.16.1
identify==2.6.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.1
mypy==1.11.2
mypy-extensions==1.0.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas-stubs==2.2.3.241009
platformdirs==4.3.6
pre_commit==4.0.1
prompt-toolkit==3.0.36
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.2
PyYAML==6.0.2
questionary==2.0.1
regex==2024.9.11
requests==2.32.3
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
termcolor==2.5.0
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.66.5
types-pytz==2024.2.0.20241003
types-PyYAML==6.0.12.20240917
types-simplejson==3.19.0.20240801
types-tzlocal==5.1.0.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
wcwidth==0.2.13
